% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codaSeq.effect.r
\name{codaSeq.effect}
\alias{codaSeq.effect}
\title{Estimate effect size between two distributions}
\usage{
codaSeq.effect(x, conds, corrected = FALSE)
}
\arguments{
\item{x}{A numerical matrix with samples by column}

\item{conds}{A vector denoting group membership}

\item{corrected}{Whether to scale to Cohen's d or not,
default is FALSE}
}
\value{
returns a vector of effect sizes
}
\description{
Calculates relatively robust measures of standardized mean difference (effect) and dispersion from a matrix or dataframe of numbers. The output value is smaller thant Cohen's d by a factor of 1.418 when comparing two Normal distributions. There is an option to normalize the effect size by that factor.
}
\examples{
# make a synthetic dataset
d <- c(rnorm(100,0,1), rnorm(100,2,1))
e <- c(rnorm(100,2,1), rnorm(100,0,1))
 de <- rbind(d,e)
conds <- c(rep("A", 100), rep("B",100))
# values should be approximately -2 and 2
codaSeq.effect(de, conds, corrected=TRUE)
}
