\name{hmpgenera}
\docType{data}
\alias{hmpgenera}
\title{Selection-based differential sequence variant abundance dataset}
\description{
  This data set is the genus table table for 15 random op and ak samples from
  the original Human Microbiome dataset.
}
\usage{data(hmpgenera)}
\format{A dataframe of 4546 genera and 6 taxonomic levels named in lower case.
    Taxa without assignment are labelled "unk"}
\source{HMP consortium}
\references{
 NIH HMP Working Group et al (2009) Genome Res dx.doi.org/10.1101/gr.096651.109
}
\keyword{datasets}
