% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codaSeq.clr.r
\name{codaSeq.clr}
\alias{codaSeq.clr}
\title{Center log-ratio function}
\usage{
codaSeq.clr(x, IQLR = FALSE, aitch = FALSE, samples.by.row = TRUE)
}
\arguments{
\item{x}{A matrix or dataframe with samples by row or column}

\item{IQLR}{the geometric mean computed on all features if FALSE,
or on the set of features with variance between the first and
third quartile if TRUE. To be used when the data is not centered.}

\item{aitch}{returns the values drawn from a digamma distribution using
an uniformative prior. Approximates the expected value of the clr
values and is much faster than using aldex.effect}

\item{samples.by.row}{TRUE if samples by row, FALSE if samples by column}
}
\value{
returns a matrix of clr tranformed values with samples
in the rows and variables in columns
}
\description{
Returns a matrix of center log-ratio transformed data with samples
by row. Each value is equivalent to log(x/gx) where gx is
the geometric mean of the row vector X.
}
\details{
the natural logarithm is used.
}
\examples{
# get dataset from ALDEx2 package
data(selex)
# convert to clr with an uninformative prior
# use only the variables with mid-quartile variance as denominator
# output will have samples by rows
selex.clr <- codaSeq.clr(selex + 0.5, IQLR=TRUE, aitch=TRUE, samples.by.row=FALSE)
}
\seealso{
\code{\link{codaSeq.filter}},
\code{\link{codaSeq.outlier}},
\code{\link{codaSeq.rarefy}},
\code{\link{codaSeq.phi}}
}
\author{
Greg Gloor, Andrew Fernandes, Jean Macklaim, Wallace Chan
}
