% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codaSeq.rarefy.r
\name{codaSeq.rarefy}
\alias{codaSeq.rarefy}
\title{Subsample a frequency table without replacement}
\usage{
codaSeq.rarefy(x, n = 1000, samples.by.row = TRUE)
}
\arguments{
\item{x}{A matrix or dataframe containing a count table.}

\item{n}{The desired target number of reads per sample (Default: \code{1000}).}

\item{samples.by.row}{A logical value indicating if samples are in rows (\code{TRUE})
or in columns (\code{FALSE}) (Default: \code{TRUE}).}
}
\value{
Returns a matrix in the same orientation as the original with counts
per OTU reduced to the common sampling depth. This is a constant sum operation.
}
\description{
Returns a table of counts where samples have been sampled without replacement.
This is included for compatibility, but in general is a bad idea, since it
results in a loss of information, and distorts the underlying data somewhat.
}
\examples{
#' # load example HMP data from CoDaSeq package
data("ak_op")  # feature table: 4347 OTU x 30 samples (15x ak, 15x op)

# calculate total number of reads in each sample (all are >1000)
colSums(ak_op)

# rarefy read table to a depth of 1000 reads
codaSeq.rarefy(ak_op, n= 1000, samples.by.row= FALSE)


}
\author{
Greg Gloor, Jean Macklaim, Wallace Chan
}
